Rebuild the plugin scaffolding
I ran `yo joplin` as I found that the new way of calling scripts `joplin.contentScripts` was not found in the actual version.

Tried to leave the `manifest.json` as close as possible to the original one. I decided to move plugin version to 2.0 as `app_min_version` has moved from 1.4 to 2.2